handler: Fix vring initialization logic
Since it's the guest decision to rely on polling or interrupts to be
notified about used descriptors in the used ring, we can't expect an
EventFd to be set through SET_VRING_CALL if the guest chose the polling
method.

On the other hand, we always expect an EventFd to be provided through
SET_VRING_KICK since the current way of handling new descriptors is
exclusively by receiving an event through an eventfd (no polling mode
has been implemented).

That's why a vring should be initialized based on its state and if the
EventFd related to VRING_KICK has been set.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>